pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["tools", "sessions", "skills", "canvas", "channels"]

[project]
name = "ollama-mlflow-demo"
version = "0.1.0"
description = "AgentCraft - AI Agent with LLM and tools"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.10.0",  # Cron scheduling
    "ddgs>=9.14.2",
    "docker>=7.0.0",
    "duckduckgo-search>=8.1.1",
    "mlflow>=3.12.0",
    "openai>=2.34.0",
    "pyinstaller>=6.20.0",
    "python-dotenv>=1.0.0",
    "python-pptx>=1.0.2",
    "pywebview>=5.0.0",  # Desktop App
    "pyyaml>=6.0",
    "qrcode>=8.2",  # 二维码生成
    "rich>=13.0.0",
    "tiktoken>=0.7.0",
    "wechatpy>=1.8.18",  # 企业微信官方API
]

[project.scripts]
agentcraft = "cli:main"
agentcraft-desktop = "desktop.mac_app:run_mac_desktop"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "py2app>=0.28.0",  # Mac app bundler
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-m 'not integration'"
markers = [
    "integration: marks tests that require a running LLM (Ollama) and valid API key",
]

[tool.uv.workspace]
members = [
    "test-pdf",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "transformers>=5.8.0",
]
